feat: improve share button to be more flexible
